requirements.txt
streamlit
google-generativeai
python-dotenv
opencv-python
numpy
fastapi
uvicorn
